refactor: split core.rs and render.rs to stay under 500-line CI limit
Extract material/skinned methods to core_materials.rs and core_skinned.rs.
Extract postprocess/fxaa helpers to render_helpers.rs.
